Sharing update_dict used in nautilus and zimit tests
