+++ dispatcher/backend/src/tests/integration/routes/schedules/test_zimit.py
@@ -2,27 +2,9 @@ from collections import namedtuple
 from typing import List
 
 import pytest
+from utils_for_tests import update_dict
 
 from common import constants
-
-
-def update_dict(dict: dict, key_path: str, new_value: any):
-    """Update a nested key value in a dictionary
-
-    E.g if key_path is 'key1.subkey2', then dict['key1']['subkey2'] will be set"""
-
-    # Split the key path into individual keys
-    keys = key_path.split(".")
-
-    # Initialize a reference to the nested dictionary
-    current_dict = dict
-
-    # Navigate through the nested structure
-    for key in keys[:-1]:
-        current_dict = current_dict[key]
-
-    # Update the value using the last key
-    current_dict[keys[-1]] = new_value
 
 
 class TestZimit:
