+++ dispatcher/backend/src/tests/utils_for_tests.py
@@ -28,3 +28,22 @@ def patch_dict(data, patch):
             # If the key is not present in the original dictionary, set it with the
             # patch value
             data[key] = patch_value
+
+
+def update_dict(dict: dict, key_path: str, new_value: any):
+    """Update a nested key value in a dictionary
+
+    E.g if key_path is 'key1.subkey2', then dict['key1']['subkey2'] will be set"""
+
+    # Split the key path into individual keys
+    keys = key_path.split(".")
+
+    # Initialize a reference to the nested dictionary
+    current_dict = dict
+
+    # Navigate through the nested structure
+    for key in keys[:-1]:
+        current_dict = current_dict[key]
+
+    # Update the value using the last key
+    current_dict[keys[-1]] = new_value
